Fix for dropping and creating tables before insert.
That's been bugging me for a while.

It probably shouldn't need to be called explicitely, although the
option is nice.

Also I'm not 100% sure if this is the best way to handle setting
the sqlalchemy objects, but the globals weren't being set correctly
the way I was importing and initializing them before.