Add part of speach table.
Hardcoded it into the parser because I don't think it's possible
to easily read out of the xml file (at least not with lxml). I
doubt these will change that much at this point, but you never know.

The next step is to alter the parser so this information can be
linked to the sense elment.