Redid the parser to not use sqlalchemy for inserts.
The files (while not yet quite complete) insert all of this data
in seconds rather than hours.

Perhaps there are things that can be done with sqlalchemy to speed
things up. Although I doubt it will ever match a bulk insert.

That's not a huge priority, I started using it mostly to make
querying easier not necessarily insertions.