Added kanji/kana to entry join tables.
I really need a way to get ids at this point. It seems that querying
the table as is right now is noticably slower than before. Or maybe
it's just my imagination.

Working on the pos now. Simplified the error checking before
thinking that it should be a separate commit.