Add gloss entry join table.
Thinking about it, I'm not entirely sure that the sense element
was entirely necessary. I couldn't find a unique way to identify
them aside from the entire element, so it made sense to just move
the part of speach into the gloss and run with that.

Who knows, I may be completely wrong about it.

It is becoming increasingly obvious that I need to references ids
somehow, this join table is noticably slow, so figuring that out
will be the first step to try fixing it.