Add ids for join table.
Incrementing my own ids as a solution.

I don't like how I figured out how to do this, and I certainly
don't like that it's valid sqlite syntax...but such is life.

The sub selects weren't working, so I guess for now this works.
Then again I was trying to cut out everything I could to increse
the speed of the conversion.