Tried setting up a couple specific dictionaries.
Using those instead to track joins. Not sure it's really helping.
Things seem to be missing from the final listing, but I can't
quite seem to track them down.

I wonder if it might be easier to keep a list of joins, and read
the ids back out from the table? Then again the big problem seems
to be a lack of kana kanji link and spead of reading the list back
out more than anything else.