Add "warehouse" table.
Basically a simple collection of all of the data, un-normalized,
but faster to search and display.

The sad fact of the matter is that it's probably exactly what I
want, aside from the glosses not containing all of their information.

Well I should say I do still want to have a normalized table to
do random querying against, but for the other immediate tools I want to
build this will probably work perfectly (i.e. it's fast).