Add unit tests for the parser.
Also required updating the parse to return a list of Entry objects. Rather
than the combination of items from before. Probably a better decision overall
but this means that the database insert no longer works.

Also created a gloss class due to the annoyence of dealing with tuples.

Finall moved the JMdict file to the main directory as it tends to be easier
to type when testing. So added it to the ignore file.