Add alternate warehouse implementation and list values.
Only read the values back out as a test to see how it looks. I am lazily
trying to match the original EDICT format.

I don't like the original warehouse implementation because it relies on comma
separated values, which *is* the point, but it smells bad.

I'm not entirely sold on the other two attempts and the moment, but it does
depend on the ultimate goal, which is searching for various text. Knowing the
limitations on the data technically the original warehouse could be perfect.