Update to improve the project structure.
I guess It could be argued whether this is actually better. The main reason
behind it is that I wanted to add support for things like mongo, and revisit
using sqlalchemy now that a few versions have come and gone.

It still lacks some Python niceties, such as package information, but it
should be an improvement moving forward.