Add kanjidc2 parser.
Yes, this has always been part of the plan to include kanji parsing, although
with how infrequently I update this it's taken a while to get to this point...
especially becasue the jmdict part is nowhere near done.

Anyway, I basically copied the xml structure as objects. I suspect there are
waus to automatically go from xml to objects, although with the file sizes
this work may make sense. Either way I want to go over it again later and
improve upon this if possible, so it isn't a one-to-one representation.
(Unless that proves to be the best choice, anyway).

Still todo:
- Add this to the import options in main.
- Add sql/whatever to save the data.
- Complete more tests.