Update to join entries and other bulk tables.
It works, but compared to the other bulk inserts it's incredibly slow
and fairly time consuming. There are probably ways to minimize that
more, but I wanted to get it up and running before messing around
futher.

I updated the tests to test a given entry and added some others that had
shown some exceptions to the rules when originally writing the code.
(Technically it's probably all laid out in the xml definition, but if I
even read through the whole thing then it's been a while, it would be
worth double checking).

I also updated the warehouse to use a different separator as a comma
could be used inside a given bit of text.